+++ toot/commands.py
@@ -100,6 +100,7 @@ def post(app, user, args):
         sensitive=args.sensitive,
         spoiler_text=args.spoiler_text,
         in_reply_to_id=args.reply_to,
+        language=args.language,
     )
 
     print_out("Toot posted: <green>{}</green>".format(response.get('url')))